django_backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
blinker==1.9.0
click==8.1.8
Django==5.1.6
django-ninja==1.3.0
extension==1.0.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
Markdown==3.7
MarkupSafe==3.0.2
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
Werkzeug==3.1.3
